Separate fhir and opensrp packages
